+++ app/Http/Controllers/ProviderTypeController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Customer;
+use App\Provider;
 use App\ProviderType;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
@@ -23,7 +24,15 @@ class ProviderTypeController extends Controller
         $validator = Validator::make($request->all(), [
             'name' => 'required|min:3',
             'description'=>'required|min:3',
+        ], [
+            'name.required' => 'Es necesario ingresar un nombre para el tipo de proveedor.',
+            'name.min' => 'El nombre debe tener al menos 3 caracteres.',
+            'description.required'=> 'Es necesario ingresar una descripcion para el tipo de proveedor.',
+            'description.min'=> 'La descripción debe tener al menos 3 caracteres.',
         ]);
+
+        if ($validator->fails())
+            return back()->withErrors($validator)->withInput();
 
         $provider_type = ProviderType::find( $request->get('id') );
         $provider_type->name = $request->get('name');
@@ -60,9 +69,20 @@ class ProviderTypeController extends Controller
     public function delete( Request $request )
     {
         $validator = Validator::make($request->all(), [
-            'id' => 'exists:provider,id'
+            'id' => 'exists:provider_types,id'
         ]);
-        //$type_id = Customer::
+        $provider = Provider::where( 'provider_type_id', $request->get('id') )->first();
+
+        if ($validator->fails() OR $provider != null)
+        {
+            $data['errors'] = $validator->errors();
+            if( $provider != null )
+                $data['errors']->add("id", "No puede eliminar el tipo de proveedor seleccionado, porque existen proveedores asignados a ese tipo.");
+
+            return redirect('/proveedores/tipos')
+                ->withInput($request->all())
+                ->with($data);
+        }
         $provider_type = ProviderType::find( $request->get('id') );
         $provider_type->delete();
